Fix: Facade 패턴을 이용해 순환참조 문제 해결
