+++ src/main/java/org/gdsccau/team5/safebridge/domain/team/controller/TeamController.java
@@ -5,6 +5,7 @@ import org.gdsccau.team5.safebridge.common.code.success.TeamSuccessCode;
 import org.gdsccau.team5.safebridge.common.response.ApiResponse;
 import org.gdsccau.team5.safebridge.domain.team.dto.request.TeamRequestDto.TeamCreateRequestDto;
 import org.gdsccau.team5.safebridge.domain.team.dto.response.TeamResponseDto.TeamDataDto;
+import org.gdsccau.team5.safebridge.domain.team.facade.TeamFacade;
 import org.gdsccau.team5.safebridge.domain.team.service.TeamService;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -15,34 +16,34 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/team")
+@RequestMapping("/teams")
 public class TeamController {
 
-    private final TeamService teamService;
+    private final TeamFacade teamFacade;
 
     @PostMapping("")
     public ApiResponse<Void> createTeam(@RequestBody final TeamCreateRequestDto requestDto) {
-        teamService.createTeam(requestDto);
+        teamFacade.createTeam(requestDto);
         return ApiResponse.onSuccess(TeamSuccessCode.CREATE_TEAM);
     }
 
     @DeleteMapping("/{teamId}")
     public ApiResponse<Void> deleteTeam(@PathVariable final Long teamId) {
-        teamService.deleteTeam(teamId);
+        teamFacade.deleteTeam(teamId);
         return ApiResponse.onSuccess(TeamSuccessCode.DELETE_TEAM);
     }
 
-    @PostMapping("/{teamId}/join/{userId}")
+    @PostMapping("/{teamId}/users/{userId}/join")
     public ApiResponse<TeamDataDto> joinTeam(@PathVariable(name = "teamId") final Long teamId,
                                              @PathVariable(name = "userId") final Long userId) {
-        TeamDataDto teamDataDto = teamService.joinTeam(teamId, userId);
+        TeamDataDto teamDataDto = teamFacade.joinTeam(teamId, userId);
         return ApiResponse.onSuccess(TeamSuccessCode.JOIN_TEAM, teamDataDto);
     }
 
-    @PostMapping("/{teamId}/leave/{userId}")
+    @PostMapping("/{teamId}/users/{userId}/leave")
     public ApiResponse<Void> leaveTeam(@PathVariable(name = "teamId") final Long teamId,
                                        @PathVariable(name = "userId") final Long userId) {
-        teamService.leaveTeam(teamId, userId);
+        teamFacade.leaveTeam(teamId, userId);
         return ApiResponse.onSuccess(TeamSuccessCode.LEAVE_TEAM);
     }
 }
